Extract babel plugin
