+++ build-system/babel-plugins/babel-plugin-transform-stop-property-mangling/test/fixtures/transform/class-method/options.json
@@ -1,0 +1,3 @@
+{
+  "plugins": ["../../../../../babel-plugin-transform-stop-property-mangling"]
+}
